initial integration between api and app

+++ GroceryCartAPI/Models/CartLinetem.cs
@@ -4,7 +4,7 @@ public class CartLinetem
 {
     public GroceryItem? GroceryItem { get; set; }
     public int Quantity { get; set; }
-    public int BasePrice { get; set; }
+    public decimal BasePrice { get; set; }
     public decimal Tax { get; set; }
     public decimal Price => Quantity * (BasePrice + Tax);
 }
